refactor: splits backend into its components
