+++ tests/scripts/generate_golden.py
@@ -19,7 +19,7 @@ from pathlib import Path
 sys.path.insert(0, str(Path(__file__).resolve().parent.parent / "src"))
 
 from backend.replay import ReplayScraper  # noqa: E402
-from backend.z_scraper import ZScraper  # noqa: E402
+from orchestrator import ZScraper  # noqa: E402
 
 GOLDEN_SCHEMA = "zview-golden/1"
